Fix solver log file generation issue when mesh generation fails

+++ foamCase/batchRun.py
@@ -112,8 +112,16 @@ def store_log_files(name):
     os.system(f'cp log.snappyHexMesh ../logs/log.snappyHexMesh_' + name)
     os.system(f'cp log.checkMesh_snapping ../logs/log.checkMesh_snapping_' + name)
     os.system(f'cp log.checkMesh_layers ../logs/log.checkMesh_layers_' + name)
-    os.system(f'cp solverCase/log.simpleFoam ../logs/log.simpleFoam_' + name)
-    os.system(f'cp solverCase/postProcessing/minMaxMagnitude\(\)/0/fieldMinMax.dat ../logs/log.fieldMinMax_' + name)
+
+    # Create empty log files if no files were created, to make postprocessing easy
+    if not os.path.isfile("solverCase/log.simpleFoam"):
+        open("../logs/log.simpleFoam_" + name, "a").close()
+    else:
+        os.system(f'cp solverCase/log.simpleFoam ../logs/log.simpleFoam_' + name)
+    if not os.path.isfile("solverCase/postProcessing/minMaxMagnitude()/0/fieldMinMax.dat"):
+        open("../logs/log.fieldMinMax_" + name, "a").close()
+    else:
+        os.system(f'cp solverCase/postProcessing/minMaxMagnitude\(\)/0/fieldMinMax.dat ../logs/log.fieldMinMax_' + name)
 
 
 snappyDict = './system/snappyHexMeshDict'
